git-svn-id: https://xp-dev.com/svn/GoldenBoot474@104 7f6f3b01-10c3-48ce-9242-e0f477912f09

+++ robotAlgorithm/robotAlgorithm/GLOBALS.h
@@ -31,7 +31,7 @@ struct Robot {
 struct Ball {
 	double x;
 	double y;
-	double rad;
+	int rad;
 };
 
 struct Obstacle {
